Moves DXF ops to the plugin gama.extension.dxf
Complete, including the library (kabeja), the models..
