+++ gama.dependencies/META-INF/MANIFEST.MF
@@ -232,7 +232,6 @@ Export-Package: COM.ibm.db2.app,
  com.sun.media.jai.tilecodec,
  com.sun.media.jai.util,
  com.sun.media.jai.widget,
- dk.abj.svg.action,
  edu.umd.cs.findbugs.annotations,
  gama.dependencies,
  gama.dependencies.geojson,
@@ -1292,41 +1291,6 @@ Export-Package: COM.ibm.db2.app,
  org.joda.time.tz,
  org.json.simple,
  org.json.simple.parser,
- org.kabeja,
- org.kabeja.batik.tools,
- org.kabeja.dxf,
- org.kabeja.dxf.helpers,
- org.kabeja.dxf.objects,
- org.kabeja.io,
- org.kabeja.math,
- org.kabeja.parser,
- org.kabeja.parser.dxf,
- org.kabeja.parser.dxf.filter,
- org.kabeja.parser.entities,
- org.kabeja.parser.header,
- org.kabeja.parser.objects,
- org.kabeja.parser.table,
- org.kabeja.processing,
- org.kabeja.processing.event,
- org.kabeja.processing.helper,
- org.kabeja.processing.scripting,
- org.kabeja.processing.scripting.impl,
- org.kabeja.svg,
- org.kabeja.svg.action,
- org.kabeja.svg.dimension,
- org.kabeja.svg.generators,
- org.kabeja.svg.tools,
- org.kabeja.svg.ui,
- org.kabeja.tools,
- org.kabeja.ui,
- org.kabeja.ui.event,
- org.kabeja.ui.impl,
- org.kabeja.ui.model,
- org.kabeja.ui.model.adapter,
- org.kabeja.ui.viewer,
- org.kabeja.ui.xml,
- org.kabeja.xml,
- org.kabeja.xslt,
  org.locationtech.jts,
  org.locationtech.jts.algorithm,
  org.locationtech.jts.algorithm.construct,
@@ -1743,7 +1707,6 @@ Bundle-ClassPath: .,
  libs/geotools 33.4/spotbugs-annotations-4.0.2.jar,
  libs/geotools 33.4/systems-common-2.1.jar,
  libs/geotools 33.4/xml-commons-resolver-1.2.jar,
- libs/kabeja 0.5/kabeja-0.5.0.jar,
  libs/jai-imageio-core 1.4.0/jai-imageio-core-1.4.0.jar,
  libs/geotools 33.4/jai_codec-1.1.3.jar,
  libs/jai_core 1.1.3/jai_core-1.1.3.jar,
