+++ gama.extension.dxf/META-INF/MANIFEST.MF
@@ -1,0 +1,17 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Gama DXF Extension
+Bundle-SymbolicName: gama.extension.dxf;singleton:=true
+Bundle-Version: 0.0.0.qualifier
+Export-Package: gama.extension.dxf
+Bundle-Vendor: UMMISCO
+Require-Bundle: gama.core,
+ gama.dependencies,
+ org.eclipse.core.resources;bundle-version="[3.23.0,4.0.0)",
+ org.eclipse.core.runtime;bundle-version="3.34.200"
+Bundle-ClassPath: .,
+ libs/kabeja 0.5/kabeja-0.5.0.jar
+Bundle-ActivationPolicy: lazy
+Bundle-RequiredExecutionEnvironment: JavaSE-25
+Automatic-Module-Name: gama.extension.dxf
+
