+++ gama.extension.dxf/build.properties
@@ -1,0 +1,10 @@
+source.. = src/,\
+           gaml/
+output.. = bin/
+jre.compilation.profile = JavaSE-25
+bin.includes = .,\
+               plugin.xml,\
+               META-INF/,\
+               models/,\
+               libs/kabeja 0.5/kabeja-0.5.0.jar
+javacDefaultEncoding.. = UTF-8
