+++ gama.parent/pom.xml
@@ -136,6 +136,8 @@
 		<module>../gama.extension.batch</module>
 		<module>../gama.extension.fipa</module>
 		<module>../gama.extension.osm</module>
+		<module>../gama.extension.dxf</module>
+		<module>../gama.extension.kml</module>		
 		<module>../gama.extension.dataframe</module>
 		<module>../gama.extension.sound</module>
 		<module>../gama.extension.stats</module>
